final/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "contextforge"
version = "0.1.0"
description = "Production-ready RAG service with FastAPI, Groq, and SPLADE."
authors = [
  { name = "Vijay", email = "vijay@example.com" },
]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "pydantic>=2.0.0",
    "psutil>=5.9.0",
    "langchain>=0.1.0",
    "langchain-community>=0.0.10",
    "langchain-core>=0.1.10",
    "langchain-experimental>=0.0.47",
    "langchain-groq>=0.0.1",
    "chromadb>=0.4.22",
    "sentence-transformers>=2.3.1",
    "transformers>=4.37.0",
    "torch>=2.1.0",
    "rich>=13.7.0",
    "python-dotenv>=1.0.0",
    "pymupdf>=1.23.8",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "black",
    "isort",
    "flake8",
]

[tool.setuptools.packages.find]
where = ["src"]
